+++ src/factory.ts
@@ -1,4 +1,5 @@
 /**
  * factory
  */
-export * from '@pecorino/factory';
+// export * from '@pecorino/factory';
+export * from '@chevre/factory';